:recycle: Refactor: component/category 폴더 분리
- refactor: 탭 매거진 container 폴더로 이동
- feat: 매거진 페이지 라우팅 처리
