+++ src/components/category/Select/SelectCategoryWithContent.tsx
@@ -1,5 +1,5 @@
-import CategoryContentList from '@/components/category/CategoryContentList';
-import SelectCategory from '@/components/category/SelectCategory';
+import CategoryContentList from '@/components/category/List/CategoryContentList';
+import SelectCategory from '@/components/category/Select/SelectCategory';
 
 // TODO query 요청 시 Photo가 렌더링이 되지 않아 토스트 팝업이 버벅여 보이는 문제
 const SelectCategoryWithContent = () => {
