+++ src/components/common/Modal/Confirm.tsx
@@ -6,18 +6,17 @@ interface ConfirmProps {
   title: string;
   onReject?: () => void;
   onSuccess?: () => void;
+  description?: string;
 }
-const Confirm = ({ title, onReject, onSuccess }: ConfirmProps) => {
+const Confirm = ({ title, onReject, onSuccess, description }: ConfirmProps) => {
   return (
     <div
       css={css`
-        width: 86vw;
         padding: 32px 38px;
         display: flex;
         flex-direction: column;
         justify-content: center;
         align-items: center;
-        gap: 38px;
       `}
     >
       <span
@@ -31,8 +30,23 @@ const Confirm = ({ title, onReject, onSuccess }: ConfirmProps) => {
       >
         {title}
       </span>
+      {description ? (
+        <span
+          css={(theme) =>
+            css`
+              margin-top: 4px;
+              ${theme.font.R_BODY_13};
+              line-height: 160%;
+              color: ${theme.color.gray07};
+            `
+          }
+        >
+          {description}
+        </span>
+      ) : null}
       <div
         css={css`
+          margin-top: 38px;
           width: 100%;
           display: flex;
           justify-content: center;
