+++ src/containers/magazine/Magazine.stories.tsx
@@ -3,6 +3,7 @@ import React from 'react';
 
 import MagazineCreateContainer from '@/containers/magazine/MagazineCreateContainer';
 import PageViewContainer from '@/containers/magazine/PageViewContainer';
+import TabMagazineViewContainer from '@/containers/magazine/TabMagazineViewContainer';
 
 export default {
   title: 'containers/magazine',
@@ -13,3 +14,6 @@ export const MagazineCreate: ComponentStory<typeof MagazineCreateContainer> = (a
   <MagazineCreateContainer {...args}>MagazineCreateContainer</MagazineCreateContainer>
 );
 export const PageView: ComponentStory<typeof PageViewContainer> = (args) => <PageViewContainer {...args} />;
+export const TabMagazine: ComponentStory<typeof TabMagazineViewContainer> = (args) => (
+  <TabMagazineViewContainer {...args} />
+);
