+++ src/containers/magazine/TabMagazineViewContainer.tsx
